account template and better indexer
Updated indexer to add a go back if you pass on a true boolean !! It does not handle the output well if there is no if check AFTER the MenuIndexer Method !!
Added AccountTemplate that has default account names and interest rates hard coded. thanks to juiceghost! had the idea of having it in the database but we decided to hard code it instead.